dynamic level giant text #2

+++ Assets/Operational/Script/script_cannon_settings.cs
@@ -17,7 +17,7 @@ public class script_cannon_settings {
     public float m_heigh_bottom;
     public float m_width_max;
     [System.NonSerialized] public float m_width_min;
-    public  float m_max_shots;
+    public  int m_max_shots;
     public  bool m_speedy = false;
     public  hot m_hot;
